Remove deprecated mapping constants (#6914)
Remove the deprecated mappings as scheduled.

Fixes https://github.com/onnx/onnx/issues/4554

---------
